# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI-Powered Travel Management Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /ai/chat:
    post:
      operationId: aiChat
      tags: [ai]
      summary: Send message to AI travel assistant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiChatBody"
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiChatResponse"
...
  /bookings:
    get:
      operationId: listBookings
      tags: [bookings]
      summary: List user bookings
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [confirmed, pending, cancelled, completed]
        - name: type
          in: query
          schema:
            type: string
            enum: [flight, hotel, activity, car_rental, package]
      responses:
        "200":
          description: List of bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingBody"
      responses:
        "201":
          description: Created booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}:
    get:
      operationId: getBooking
      tags: [bookings]
      summary: Get booking detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingDetail"
    patch:
      operationId: updateBooking
      tags: [bookings]
      summary: Update booking (modify)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBookingBody"
      responses:
        "200":
          description: Updated booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}/cancel:
    post:
      operationId: cancelBooking
      tags: [bookings]
      summary: Cancel a booking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CancelBookingBody"
      responses:
        "200":
          description: Cancellation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CancellationResult"

  # ==================== EXPENSES ====================
...
  /corporate/approvals:
    get:
      operationId: listApprovals
      tags: [corporate]
      summary: List pending approvals
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected]
      responses:
        "200":
          description: Approvals list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Approval"
  /corporate/approvals/{id}/approve:
    post:
      operationId: approveRequest
      tags: [corporate]
      summary: Approve a travel request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Approval"
  /corporate/approvals/{id}/reject:
    post:
      operationId: rejectRequest
      tags: [corporate]
      summary: Reject a travel request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectRequestBody"
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Approval"
  /corporate/policy:
    get:
      operationId: getTravelPolicy
      tags: [corporate]
      summary: Get company travel policy
      responses:
        "200":
          description: Travel policy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TravelPolicy"
    patch:
      operationId: updateTravelPolicy
      tags: [corporate]
      summary: Update company travel policy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTravelPolicyBody"
      responses:
        "200":
          description: Updated policy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TravelPolicy"
  /corporate/spend-report:
    get:
      operationId: getCorporateSpendReport
      tags: [corporate]
      summary: Get corporate spend analytics
      parameters:
        - name: period
          in: query
          schema:
            type: string
        - name: groupBy
          in: query
          schema:
            type: string
            enum: [department, traveler, destination, category]
      responses:
        "200":
          description: Spend report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpendReport"

  # ==================== NOTIFICATIONS ====================
...
components:
  schemas:
...
    AiChatBody:
      type: object
      properties:
        message:
          type: string
        sessionId:
          type: ["string", "null"]
        context:
          type: ["string", "null"]
      required:
        - message
    AiChatResponse:
      type: object
      properties:
        reply:
          type: string
        sessionId:
          type: string
        suggestions:
          type: array
          items:
            type: string
        type:
          type: string
          enum: [general, booking, planning, alert, info]
      required:
        - reply
        - sessionId
        - suggestions
        - type
...
    Booking:
      type: object
      properties:
        id:
          type: integer
        tripId:
          type: ["integer", "null"]
        type:
          type: string
          enum: [flight, hotel, activity, car_rental, package]
        status:
          type: string
          enum: [confirmed, pending, cancelled, completed]
        reference:
          type: string
        providerName:
          type: string
        description:
          type: string
        startDate:
          type: string
        endDate:
          type: ["string", "null"]
        amount:
          type: number
        currency:
          type: string
        passengerCount:
          type: integer
        isRefundable:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - tripId
        - type
        - status
        - reference
        - providerName
        - description
        - startDate
        - endDate
        - amount
        - currency
        - passengerCount
        - isRefundable
        - createdAt
    BookingDetail:
      type: object
      properties:
        id:
          type: integer
        tripId:
          type: ["integer", "null"]
        type:
          type: string
        status:
          type: string
        reference:
          type: string
        providerName:
          type: string
        description:
          type: string
        startDate:
          type: string
        endDate:
          type: ["string", "null"]
        amount:
          type: number
        currency:
          type: string
        passengerCount:
          type: integer
        isRefundable:
          type: boolean
        passengers:
          type: array
          items:
            type: string
        metadata:
          type: object
          properties:
            origin:
              type: ["string", "null"]
            destination:
              type: ["string", "null"]
            flightNumber:
              type: ["string", "null"]
            seatNumbers:
              type: ["string", "null"]
            hotelAddress:
              type: ["string", "null"]
            confirmationDetails:
              type: ["string", "null"]
          required:
            - origin
            - destination
            - flightNumber
            - seatNumbers
            - hotelAddress
            - confirmationDetails
        createdAt:
          type: string
      required:
        - id
        - tripId
        - type
        - status
        - reference
        - providerName
        - description
        - startDate
        - endDate
        - amount
        - currency
        - passengerCount
        - isRefundable
        - passengers
        - metadata
        - createdAt
    CreateBookingBody:
      type: object
      properties:
        tripId:
          type: ["integer", "null"]
        type:
          type: string
        providerName:
          type: string
        description:
          type: string
        startDate:
          type: string
        endDate:
          type: ["string", "null"]
        amount:
          type: number
        currency:
          type: string
        passengerCount:
          type: integer
        isRefundable:
          type: boolean
      required:
        - type
        - providerName
        - description
        - startDate
        - amount
        - currency
        - passengerCount
        - isRefundable
    UpdateBookingBody:
      type: object
      properties:
        status:
          type: string
        startDate:
          type: string
        endDate:
          type: ["string", "null"]
    CancelBookingBody:
      type: object
      properties:
        reason:
          type: string
      required:
        - reason
    CancellationResult:
      type: object
      properties:
        success:
          type: boolean
        refundAmount:
          type: number
        refundTimeline:
          type: string
        message:
          type: string
      required:
        - success
        - refundAmount
        - refundTimeline
        - message
...
    Approval:
      type: object
      properties:
        id:
          type: integer
        travelerName:
          type: string
        travelerEmail:
          type: string
        department:
          type: string
        tripDescription:
          type: string
        destination:
          type: string
        travelDates:
          type: string
        totalAmount:
          type: number
        currency:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected]
        reason:
          type: ["string", "null"]
        submittedAt:
          type: string
        reviewedAt:
          type: ["string", "null"]
        policyCompliant:
          type: boolean
        policyNotes:
          type: ["string", "null"]
      required:
        - id
        - travelerName
        - travelerEmail
        - department
        - tripDescription
        - destination
        - travelDates
        - totalAmount
        - currency
        - status
        - reason
        - submittedAt
        - reviewedAt
        - policyCompliant
        - policyNotes
    RejectRequestBody:
      type: object
      properties:
        reason:
          type: string
      required:
        - reason
    TravelPolicy:
      type: object
      properties:
        id:
          type: integer
        maxHotelRatePerNight:
          type: number
        maxFlightClassShortHaul:
          type: string
        maxFlightClassLongHaul:
          type: string
        advanceBookingDays:
          type: integer
        requiresApprovalAbove:
          type: number
        perDiemRate:
          type: number
        currency:
          type: string
        preferredAirlines:
          type: array
          items:
            type: string
        preferredHotelChains:
          type: array
          items:
            type: string
        updatedAt:
          type: string
      required:
        - id
        - maxHotelRatePerNight
        - maxFlightClassShortHaul
        - maxFlightClassLongHaul
        - advanceBookingDays
        - requiresApprovalAbove
        - perDiemRate
        - currency
        - preferredAirlines
        - preferredHotelChains
        - updatedAt
    UpdateTravelPolicyBody:
      type: object
      properties:
        maxHotelRatePerNight:
          type: number
        maxFlightClassShortHaul:
          type: string
        maxFlightClassLongHaul:
          type: string
        advanceBookingDays:
          type: integer
        requiresApprovalAbove:
          type: number
        perDiemRate:
          type: number
        preferredAirlines:
          type: array
          items:
            type: string
        preferredHotelChains:
          type: array
          items:
            type: string
    SpendReport:
      type: object
      properties:
        period:
          type: string
        totalSpend:
          type: number
        currency:
          type: string
        groupedData:
          type: array
          items:
            $ref: "#/components/schemas/SpendGroup"
        topCategories:
          type: array
          items:
            $ref: "#/components/schemas/SpendCategory"
      required:
        - period
        - totalSpend
        - currency
        - groupedData
        - topCategories
    SpendGroup:
      type: object
      properties:
        label:
          type: string
        amount:
          type: number
        count:
          type: integer
        compliance:
          type: number
      required:
        - label
        - amount
        - count
        - compliance
    SpendCategory:
      type: object
      properties:
        category:
          type: string
        amount:
          type: number
        percentage:
          type: number
      required:
        - category
        - amount
        - percentage
...