# lib/api-zod/src/generated/types/approval.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI-Powered Travel Management Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ApprovalStatus } from "./approvalStatus";

export interface Approval {
  id: number;
  travelerName: string;
  travelerEmail: string;
  department: string;
  tripDescription: string;
  destination: string;
  travelDates: string;
  totalAmount: number;
  currency: string;
  status: ApprovalStatus;
  /** @nullable */
  reason: string | null;
  submittedAt: string;
  /** @nullable */
  reviewedAt: string | null;
  policyCompliant: boolean;
  /** @nullable */
  policyNotes: string | null;
}